+++ tests/unit/test_parser.py
@@ -32,6 +32,18 @@ class TestParser(object):
         parsed = parser.parse_args([input_path, "-m", mode])
         assert parsed.mode == mode
 
+    def test_mode_block_gappy(self, parser):
+        input_path = "my/input/file.fa"
+        mode = "block-gappy"
+        parsed = parser.parse_args([input_path, "-m", mode])
+        assert parsed.mode == mode
+
+    def test_mode_composition_bias(self, parser):
+        input_path = "my/input/file.fa"
+        mode = "composition-bias"
+        parsed = parser.parse_args([input_path, "-m", mode])
+        assert parsed.mode == mode
+
     def test_plot_trim_report_with_no_value(self, parser):
         input_path = "my/input/file.fa"
         parsed = parser.parse_args([input_path, "--plot_trim_report"])